channelData array populating update
In the original code it starts populating channelData by index 1, and later 2. Trouble is that it doesn't have index 2, only 0 and 1 as declared in line 886.
My hardware uses channel two (ADS pins 5 and 6) as electrodes input, it means I'll use channelData[1] as actual data (as it's already being used in the code).